www: Make step lock duration formatting consistent with builds
