+++ www/react-ui/src/util/DataUtils.ts
@@ -15,7 +15,7 @@
   Copyright Buildbot Team Members
 */
 
-import {Build, Builder, DataCollection, Master} from "buildbot-data-js";
+import {Build, Builder, DataCollection, Master, Step} from "buildbot-data-js";
 import {durationFormat} from "./Moment";
 
 export function hasActiveMaster(builder: Builder, masters: DataCollection<Master>) {
@@ -44,4 +44,32 @@ export function buildDurationFormatWithLocks(build: Build, now: number) {
     res += ` (locks: ${durationFormat(build.locks_duration_s!)})`;
   }
   return res;
+}
+
+export function stepDurationFormatWithLocks(step: Step, now: number) {
+  const lockDuration = step.locks_acquired_at !== null
+    ? step.locks_acquired_at - step.started_at!
+    : 0;
+
+  if (step.complete) {
+    const stepDurationText = durationFormat(step.complete_at! - step.started_at!);
+
+    if (lockDuration > 1) {
+      // Since lock delay includes general step setup overhead, then sometimes the started_at and
+      // locks_acquired_at may fall into different seconds. However, it's unlikely that step setup
+      // would take more than one second.
+      return `${stepDurationText} (locks: ${durationFormat(lockDuration)})`
+    }
+    return stepDurationText;
+  }
+
+  const ongoingStepDurationText = durationFormat(now - step.started_at!);
+  if (lockDuration > 1) {
+    // Since lock delay includes general step setup overhead, then sometimes the started_at and
+    // locks_acquired_at may fall into different seconds. However, it's unlikely that step setup
+    // would take more than one second.
+    return `${ongoingStepDurationText} (locks: ${durationFormat(lockDuration)})`
+  }
+
+  return ongoingStepDurationText;
 }
